handle missing init functions gracefully
The Go compiler is smart enough to avoid generating an init function if
there's no init work to be done. In that case, we'd fail spectacularly:

	[...]: relocation target test.tld/foo/noinit.init not defined

To fix that, always add a stub file that declares an empty init func.
This is fine, as multiple init funcs per package are allowed.