add -vet=off to the default 'go test' flags
Rarely does one want to run vet when benchmarking. More importantly,
this makes the tool faster to start running benchmarks, particularly
when the cache isn't warm. Vet will still be run as part of other 'go
test' invocations, which the developer should be using anyway.

This particularly helps the testscript tests, since each one of them
starts with an empty cache. This slashes the test time on my laptop from
~2.4s to ~0.7s.