improve support for more edge cases
If there are no non-test Go files, simply do nothing for that package.
We used to panic.

If a BenchmarkInit func already exists, the test build will fail. But
that seems fine; there's no other simple alternative. We could come up
with unique names to work around the duplicate, but if a package has a
BenchmarkInit member, it probably means it has its own way of
benchmarking init performance.

For now, let the build error surface to the user.