add -r (recursive) flag
After some initial testing, it seems like only the runtime and testing
packages cause trouble if re-initialised while the benchmark itself is
running.

It's not useful to benchmark them, anyway. All packages must depend on
runtime, and no non-test package should import testing.

This was the last TODO item, so remove the file.