refactor benchmark template
Place the benchmark at the top, since that's what's most interesting.

Remove the "no need to import?" comment, as that's clear now. We don't
need to, since the linker already has the required packages as part of
the build.

Finally, move all the "deinit" work into an appropriately named
function, and start stopping the timer when doing that potentially large
amount of work.