refactor globalsToZero to be simpler
PkgPath, Name, and TotalSize can be filled out in the parent, so do that
to deduplicate work in each of the functions. Similarly, these are all
globals, so a types.Type instead of a *types.Var is enough.

Also add a fieldByName, which finds a struct's field and its offset. The
code is simpler as a separate func, since we can use early returns.