don't reinit the os package
This was causing 'benchinit -r cmd/go' to exit abruptly for no apparent
reason. The underlying cause was a runtime panic due to a bad pointer,
but we couldn't see that as the panic message was thrown away.

Not reinitialising the os package makes the panic show up as usual. So
do that.

Add a simple regression test too, which consistently failed on the
previous version of benchinit on my machine.