get 'benchinit -r cmd/go' working
The last remaining issue turns out to be a combination of setting global
flags, along with init functions that trigger a GC run.

Not resetting the flag.CommandLine field happens to work around this
problem, for a reason still unknown. See the added TODO and tests.

It's fine to not reset that flagset though, as it's initialised via a
constructor. That happens in init already, so it already gets reset
at every init run.