add support for Go 1.13
In master, the Go compiler no longer generates an init function per
package. See https://go-review.googlesource.com/c/go/+/161337.

Adapt to that via a build flag, so that Go 1.12 and earlier continues to
work as usual. We probably want to support a couple of Go versions at
all times.