stop using a Go template for the benchmark
Instead, use a regular Go benchmark with input data as JSON.
We still need to replace the package name and in the future add Go
imports, but that's fairly easy to do at the string level.

Feeding input via JSON will be a lot easier, as we no longer need to
carefully construct Go literals to fill the data.
Moreover, we can now edit the Go code with the help of an LSP.